refactor: 과도한 interface 패턴 제거
