+++ src/main/java/com/swmStrong/demo/domain/pomodoro/service/PomodoroServiceImpl.java
@@ -19,9 +19,9 @@ import com.swmStrong.demo.domain.sessionScore.entity.SessionScore;
 import com.swmStrong.demo.domain.sessionScore.facade.SessionScoreProvider;
 import com.swmStrong.demo.domain.sessionScore.repository.SessionScoreRepository;
 import com.swmStrong.demo.domain.sessionScore.service.SessionStateManager;
+import com.swmStrong.demo.domain.user.facade.UserInfoProvider;
 import com.swmStrong.demo.infra.LLM.LLMSummaryProvider;
 import com.swmStrong.demo.domain.user.entity.User;
-import com.swmStrong.demo.domain.user.facade.UserInfoProvider;
 import com.swmStrong.demo.infra.redis.stream.RedisStreamProducer;
 import com.swmStrong.demo.infra.redis.stream.StreamConfig;
 import com.swmStrong.demo.message.dto.PomodoroPatternClassifyMessage;
