+++ src/main/java/com/swmStrong/demo/domain/user/facade/UserDeleteProvider.java
@@ -1,5 +1,18 @@
 package com.swmStrong.demo.domain.user.facade;
 
-public interface UserDeleteProvider {
-    void deleteById(String userId);
+import com.swmStrong.demo.domain.user.repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserDeleteProvider{
+
+    private final UserRepository userRepository;
+
+    public UserDeleteProvider(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+    
+    public void deleteById(String userId) {
+        userRepository.deleteById(userId);
+    }
 }
